storage/sqlite: rename to sqlite3 to match driver name/version
