+++ storage.go
@@ -10,13 +10,13 @@ import (
 	"github.com/sourcegraph/checkup/storage/postgres"
 	"github.com/sourcegraph/checkup/storage/s3"
 	"github.com/sourcegraph/checkup/storage/sql"
-	"github.com/sourcegraph/checkup/storage/sqlite"
+	"github.com/sourcegraph/checkup/storage/sqlite3"
 )
 
 func storageDecode(typeName string, config json.RawMessage) (Storage, error) {
 	switch typeName {
-	case sqlite.Type:
-		return sqlite.New(config)
+	case sqlite3.Type:
+		return sqlite3.New(config)
 	case mysql.Type:
 		return mysql.New(config)
 	case postgres.Type:
